chore: working on refactoring the dialogs in dialog.dart
