+++ lib/add_liquidity/widgets/pool_approve_button.dart
@@ -1,5 +1,5 @@
 import 'package:ax_dapp/add_liquidity/add_liquidity.dart';
-import 'package:ax_dapp/service/dialog.dart';
+import 'package:ax_dapp/service/confirmation_dialogs/custom_confirmation_dialogs.dart';
 import 'package:ax_dapp/service/failed_dialog.dart';
 import 'package:ax_dapp/service/tracking/tracking_cubit.dart';
 import 'package:ax_dapp/wallet/bloc/wallet_bloc.dart';
