+++ lib/athlete/widgets/athlete_buy_approve_button.dart
@@ -36,7 +36,7 @@ class AthleteBuyApproveButton extends StatefulWidget {
   final String walletAddress;
   final Future<void> Function() approveCallback;
   final Future<void> Function() confirmCallback;
-  final Dialog Function(BuildContext) confirmDialog;
+  final Widget confirmDialog;
 
   @override
   State<AthleteBuyApproveButton> createState() =>
@@ -118,7 +118,7 @@ class _AthleteBuyApproveButtonState extends State<AthleteBuyApproveButton> {
               showDialog<void>(
                 context: context,
                 builder: (BuildContext context) =>
-                    widget.confirmDialog(context),
+                    widget.confirmDialog,
               ).then((value) {
                 if (mounted) {
                   Navigator.pop(context);
