+++ lib/dialogs/buy/buy_dialog.dart
@@ -1,8 +1,9 @@
 import 'package:ax_dapp/athlete/athlete.dart' hide AptTypeSelectionChanged;
 import 'package:ax_dapp/dialogs/buy/bloc/buy_dialog_bloc.dart';
 import 'package:ax_dapp/scout/models/models.dart';
+import 'package:ax_dapp/service/confirmation_dialogs/custom_confirmation_dialogs.dart';
 import 'package:ax_dapp/service/controller/controller.dart';
-import 'package:ax_dapp/service/dialog.dart';
+import 'package:ax_dapp/service/custom_styles.dart';
 import 'package:ax_dapp/util/bloc_status.dart';
 import 'package:ax_dapp/util/util.dart';
 import 'package:ax_dapp/wallet/wallet.dart';
@@ -77,10 +78,10 @@ class _BuyDialogState extends State<BuyDialog> {
       child: Row(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: [
-          Text('Total Fees:', style: textStyle(Colors.grey[600]!, 15, false)),
+          Text('Total Fees:', style: textStyle(Colors.grey[600]!, 15, isBold:false)),
           Text(
             '$totalFee AX(0.3%)',
-            style: textStyle(Colors.grey[600]!, 15, false),
+            style: textStyle(Colors.grey[600]!, 15, isBold:false),
           ),
         ],
       ),
@@ -94,7 +95,7 @@ class _BuyDialogState extends State<BuyDialog> {
         children: [
           Text(
             'Balance: $balance',
-            style: textStyle(Colors.grey[600]!, 15, false),
+            style: textStyle(Colors.grey[600]!, 15, isBold:false),
           ),
         ],
       ),
@@ -108,11 +109,11 @@ class _BuyDialogState extends State<BuyDialog> {
         children: [
           Text(
             'Market Price Impact:',
-            style: textStyle(Colors.grey[600]!, 15, false),
+            style: textStyle(Colors.grey[600]!, 15, isBold:false),
           ),
           Text(
             '$marketPriceImpact %',
-            style: textStyle(Colors.grey[600]!, 15, false),
+            style: textStyle(Colors.grey[600]!, 15, isBold:false),
           ),
         ],
       ),
@@ -126,11 +127,11 @@ class _BuyDialogState extends State<BuyDialog> {
         children: [
           Text(
             'Minimum Received:',
-            style: textStyle(Colors.grey[600]!, 15, false),
+            style: textStyle(Colors.grey[600]!, 15, isBold:false),
           ),
           Text(
             '$minimumReceived APT',
-            style: textStyle(Colors.grey[600]!, 15, false),
+            style: textStyle(Colors.grey[600]!, 15, isBold:false),
           ),
         ],
       ),
@@ -144,11 +145,11 @@ class _BuyDialogState extends State<BuyDialog> {
         children: [
           Text(
             'Slippage Tolerance:',
-            style: textStyle(Colors.grey[600]!, 15, false),
+            style: textStyle(Colors.grey[600]!, 15, isBold:false),
           ),
           Text(
             '$slipageTolerance %',
-            style: textStyle(Colors.grey[600]!, 15, false),
+            style: textStyle(Colors.grey[600]!, 15, isBold:false),
           ),
         ],
       ),
@@ -205,7 +206,7 @@ class _BuyDialogState extends State<BuyDialog> {
                     children: [
                       Text(
                         'Buy ${widget.athleteName} APT',
-                        style: textStyle(Colors.white, 20, false),
+                        style: textStyle(Colors.white, 20, isBold:false),
                       ),
                       IconButton(
                         icon: const Icon(
@@ -321,7 +322,7 @@ class _BuyDialogState extends State<BuyDialog> {
                           Expanded(
                             child: Text(
                               'AX',
-                              style: textStyle(Colors.white, 15, false),
+                              style: textStyle(Colors.white, 15, isBold:false),
                             ),
                           ),
                           Container(
@@ -341,7 +342,7 @@ class _BuyDialogState extends State<BuyDialog> {
                               },
                               child: Text(
                                 'MAX',
-                                style: textStyle(Colors.grey[400]!, 9, false),
+                                style: textStyle(Colors.grey[400]!, 9, isBold:false),
                               ),
                             ),
                           ),
@@ -350,11 +351,11 @@ class _BuyDialogState extends State<BuyDialog> {
                             child: IntrinsicWidth(
                               child: TextField(
                                 controller: _aptAmountController,
-                                style: textStyle(Colors.grey[400]!, 22, false),
+                                style: textStyle(Colors.grey[400]!, 22, isBold:false),
                                 decoration: InputDecoration(
                                   hintText: '0.00',
                                   hintStyle:
-                                      textStyle(Colors.grey[400]!, 22, false),
+                                      textStyle(Colors.grey[400]!, 22, isBold:false),
                                   contentPadding:
                                       const EdgeInsets.only(left: 3),
                                   border: InputBorder.none,
@@ -457,7 +458,7 @@ class _BuyDialogState extends State<BuyDialog> {
                                 : 'Short Apt',
                             approveCallback: bloc.swapController.approve,
                             confirmCallback: bloc.swapController.swap,
-                            confirmDialog: transactionConfirmed,
+                            confirmDialog: ConfirmTransactionDialog(context: context),
                             walletAddress: formattedWalletAddress,
                           );
                         },
@@ -546,7 +547,7 @@ class Price extends StatelessWidget {
     return Row(
       mainAxisAlignment: MainAxisAlignment.spaceBetween,
       children: [
-        Text('Price:', style: textStyle(Colors.white, 15, false)),
+        Text('Price:', style: textStyle(Colors.white, 15, isBold:false)),
         BlocBuilder<BuyDialogBloc, BuyDialogState>(
           buildWhen: (previous, current) =>
               previous.aptTypeSelection != current.aptTypeSelection ||
@@ -555,7 +556,7 @@ class Price extends StatelessWidget {
               previous.shortApt != current.shortApt,
           builder: (context, state) {
             final price = state.aptBuyInfo.axPerAptPrice.toStringAsFixed(6);
-            final _textStyle = textStyle(Colors.white, 15, false);
+            final _textStyle = textStyle(Colors.white, 15, isBold:false);
             return state.aptTypeSelection.isLong
                 ? Text(
                     '$price AX per ${state.longApt.ticker} APT',
@@ -577,7 +578,7 @@ class AmountToReceive extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final _textStyle = textStyle(Colors.white, 15, false);
+    final _textStyle = textStyle(Colors.white, 15, isBold:false);
     return Row(
       mainAxisAlignment: MainAxisAlignment.spaceBetween,
       children: [
