+++ lib/scout/dialogs/redeem_dialog.dart
@@ -1,8 +1,9 @@
 import 'dart:math';
 
 import 'package:ax_dapp/scout/models/models.dart';
+import 'package:ax_dapp/service/confirmation_dialogs/custom_confirmation_dialogs.dart';
 import 'package:ax_dapp/service/controller/scout/lsp_controller.dart';
-import 'package:ax_dapp/service/dialog.dart';
+import 'package:ax_dapp/service/custom_styles.dart';
 import 'package:ax_dapp/service/failed_dialog.dart';
 import 'package:ax_dapp/service/tracking/tracking_cubit.dart';
 import 'package:ax_dapp/wallet/wallet.dart';
@@ -84,7 +85,7 @@ class _RedeemDialogState extends State<RedeemDialog> {
           Obx(
             () => Text(
               'Balance: ${longBalance.value}',
-              style: textStyle(Colors.grey[600]!, 15, false),
+              style: textStyle(Colors.grey[600]!, 15, isBold: false),
             ),
           ),
         ],
@@ -101,7 +102,7 @@ class _RedeemDialogState extends State<RedeemDialog> {
           Obx(
             () => Text(
               'Balance: ${shortBalance.value}',
-              style: textStyle(Colors.grey[600]!, 15, false),
+              style: textStyle(Colors.grey[600]!, 15, isBold: false),
             ),
           ),
         ],
@@ -116,14 +117,14 @@ class _RedeemDialogState extends State<RedeemDialog> {
         children: [
           Text(
             'You Receive: ',
-            style: textStyle(Colors.white, 15, false),
+            style: textStyle(Colors.white, 15, isBold:false),
           ),
           Obx(
             () => Row(
               children: [
                 Text(
                   '''${(lspController.redeemAmt * collateralPerPair).toStringAsFixed(6)} AX''', // 15000 is the collateral per pair
-                  style: textStyle(Colors.white, 15, false),
+                  style: textStyle(Colors.white, 15, isBold:false),
                 ),
               ],
             ),
@@ -176,7 +177,7 @@ class _RedeemDialogState extends State<RedeemDialog> {
                       padding: EdgeInsets.zero,
                       child: Text(
                         'Redeem ${widget.athlete.name} APT Pair',
-                        style: textStyle(Colors.white, 20, false),
+                        style: textStyle(Colors.white, 20, isBold:false),
                       ),
                     ),
                   ),
@@ -211,7 +212,7 @@ class _RedeemDialogState extends State<RedeemDialog> {
                       style: textStyle(
                         Colors.grey[600]!,
                         isWeb ? 14 : 12,
-                        false,
+                        isBold:false,
                       ),
                     ),
                     TextSpan(
@@ -220,7 +221,7 @@ class _RedeemDialogState extends State<RedeemDialog> {
                       style: textStyle(
                         Colors.grey[600]!,
                         isWeb ? 14 : 12,
-                        false,
+                        isBold:false,
                       ),
                     ),
                     TextSpan(
@@ -228,7 +229,7 @@ class _RedeemDialogState extends State<RedeemDialog> {
                       style: textStyle(
                         Colors.amber[400]!,
                         isWeb ? 14 : 12,
-                        false,
+                        isBold:false,
                       ),
                     ),
                   ],
@@ -244,7 +245,7 @@ class _RedeemDialogState extends State<RedeemDialog> {
                   children: [
                     Text(
                       isWeb ? 'Input APT pair:' : 'Input APT pair and amount:',
-                      style: textStyle(Colors.grey[600]!, 14, false),
+                      style: textStyle(Colors.grey[600]!, 14, isBold:false),
                     ),
                     Container(
                       margin: const EdgeInsets.only(bottom: 10),
@@ -265,7 +266,7 @@ class _RedeemDialogState extends State<RedeemDialog> {
                         },
                         child: Text(
                           'MAX',
-                          style: textStyle(Colors.grey[400]!, 9, false),
+                          style: textStyle(Colors.grey[400]!, 9, isBold:false),
                         ),
                       ),
                     ),
@@ -305,7 +306,7 @@ class _RedeemDialogState extends State<RedeemDialog> {
                           Expanded(
                             child: Text(
                               'Long APTs',
-                              style: textStyle(Colors.white, 15, false),
+                              style: textStyle(Colors.white, 15, isBold:false),
                             ),
                           ),
                           ConstrainedBox(
@@ -313,13 +314,13 @@ class _RedeemDialogState extends State<RedeemDialog> {
                             child: IntrinsicWidth(
                               child: TextField(
                                 controller: _longInputController,
-                                style: textStyle(Colors.grey[400]!, 22, false),
+                                style: textStyle(Colors.grey[400]!, 22, isBold:false),
                                 decoration: InputDecoration(
                                   hintText: '0.00',
                                   hintStyle: textStyle(
                                     Colors.grey[400]!,
                                     22,
-                                    false,
+                                    isBold:false,
                                   ),
                                   contentPadding: isWeb
                                       ? const EdgeInsets.all(9)
@@ -382,7 +383,7 @@ class _RedeemDialogState extends State<RedeemDialog> {
                           Expanded(
                             child: Text(
                               'Short APTs',
-                              style: textStyle(Colors.white, 15, false),
+                              style: textStyle(Colors.white, 15, isBold:false),
                             ),
                           ),
                           ConstrainedBox(
@@ -390,13 +391,13 @@ class _RedeemDialogState extends State<RedeemDialog> {
                             child: IntrinsicWidth(
                               child: TextField(
                                 controller: _shortInputController,
-                                style: textStyle(Colors.grey[400]!, 22, false),
+                                style: textStyle(Colors.grey[400]!, 22, isBold:false),
                                 decoration: InputDecoration(
                                   hintText: '0.00',
                                   hintStyle: textStyle(
                                     Colors.grey[400]!,
                                     22,
-                                    false,
+                                    isBold:false,
                                   ),
                                   contentPadding: isWeb
                                       ? const EdgeInsets.all(9)
@@ -453,8 +454,7 @@ class _RedeemDialogState extends State<RedeemDialog> {
                         if (result) {
                           await showDialog<void>(
                             context: context,
-                            builder: (BuildContext context) =>
-                                confirmTransaction(context, true, ''),
+                            builder: (BuildContext context) => ConfirmTransactionDialog(context: context),
                           ).then((value) {
                             final walletAddress = context
                                 .read<WalletBloc>()
@@ -485,7 +485,7 @@ class _RedeemDialogState extends State<RedeemDialog> {
                       },
                       child: Text(
                         'Confirm',
-                        style: textStyle(Colors.amber[500]!, 16, false),
+                        style: textStyle(Colors.amber[500]!, 16, isBold:false),
                       ),
                     ),
                   ),
