+++ lib/service/confirmation_dialogs/custom_confirmation_dialogs.dart
@@ -1,0 +1,3 @@
+export 'confirm_transaction_dialog.dart';
+export 'custom_confirmation_dialog.dart';
+export 'removal_confirmed_dialog.dart';
