+++ lib/wallet/widgets/dialogs/dialogs.dart
@@ -1,3 +1,5 @@
+export 'account_dialog.dart';
 export 'connect_metamask_dialog.dart';
 export 'wallet_dialog.dart';
 export 'wrong_network_dialog.dart';
+export 'your_ax_dialog.dart';
