+++ lib/wallet/widgets/wallet_account.dart
@@ -1,5 +1,5 @@
-import 'package:ax_dapp/service/dialog.dart';
-import 'package:ax_dapp/wallet/wallet.dart';
+import 'package:ax_dapp/wallet/bloc/wallet_bloc.dart';
+import 'package:ax_dapp/wallet/widgets/dialogs/dialogs.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
